Learning CS106B: Day 08-03, finished A3A-Q2

+++ assignments/Recursion/Fractals/src/fractals.cpp
@@ -82,8 +82,9 @@ void drawTreeHelper(GWindow& gw, double x, double y, double size, int order, int
     {
         int alpha = (i - 3) * 15;
         int beta = theta + alpha;
-        drawTreeHelper(gw, x + size / 2 * cos(theta / 180 * pi),
-                           y - size / 2 * sin(theta / 180 * pi),
+        // 参数一行写不下的格式？
+        drawTreeHelper(gw, x + size / 2 * cos(double(theta) / 180 * pi),
+                           y - size / 2 * sin(double(theta) / 180 * pi),
                            size / 2, order - 1, beta);
     }
 }
